refactor(estimation): extract BaseRegression from Feols
Split the result machinery every estimator shares out of `Feols` into a new
`BaseRegression` class and leave `Feols` as the OLS leaf. The move is
mechanical: construction, formula publication, within-data preparation, the
storage and capability guards, covariance dispatch, and the shared
post-estimation methods keep their bodies; only their home changes.

`get_fit()`, `_finalize_fit()`, and `_crv3_refit()` become seams: the base
declares them and `Feols` supplies the OLS solve, the goodness-of-fit and Wald
statistics, and the `feols()` jackknife refit. The difference-in-differences
placeholders move behind a `_bind_estimator_methods()` hook so the OLS leaf
keeps binding them without repeating the base constructor signature.

Result containers and reporting consumers now name `BaseRegression` instead of
the `Feols | Fepois | Feiv` union, which the base class makes the accurate
static type. `feols()` and `feglm()` keep their published return types through
a documented cast, because the runner resolves the result class from the model
registry at run time.
