+++ pyfixest/estimation/models/feols_.py
@@ -2,103 +2,60 @@ from __future__ import annotations
 
 import re
 import warnings
-from collections.abc import Mapping
-from dataclasses import replace
-from functools import partial
 from importlib import import_module
 from typing import Any, ClassVar, Literal, cast
 
 import formulaic
 import numpy as np
 import pandas as pd
-from numpy.typing import NDArray
-from scipy.sparse import csc_matrix, diags, spmatrix
-from scipy.sparse.linalg import lsqr
-from scipy.stats import chi2, f, t
+from scipy.sparse import csc_matrix, spmatrix
+from scipy.stats import t
 
-from pyfixest.core.demean import Preconditioner, WithinPreconditionerName
-from pyfixest.demeaners import AnyDemeaner, LsmrDemeaner, MapDemeaner
-from pyfixest.errors import VcovTypeNotSupportedError
-from pyfixest.estimation.api.utils import _ALL_SAMPLE, _AllSampleSentinel
 from pyfixest.estimation.capabilities import (
-    DID_METHOD_LABELS,
     DIFFERENCE_IN_DIFFERENCES,
     FIXED_EFFECTS,
     FREQUENCY_WEIGHTS,
     NON_FREQUENCY_WEIGHTS,
     WEIGHTED,
     Capabilities,
-    Feature,
-    FitFeatures,
-    require_support,
     supported,
     unless,
 )
 from pyfixest.estimation.formula import FORMULAIC_TRANSFORMS
-from pyfixest.estimation.formula import model_matrix as model_matrix_fixest
-from pyfixest.estimation.formula.formulaic_compat import (
-    materialize_model_spec_with_unseen_mask,
-)
-from pyfixest.estimation.formula.model_matrix import _ModelMatrixKey
-from pyfixest.estimation.formula.parse import Formula as FixestFormula
-from pyfixest.estimation.internals.collinearity import drop_multicollinear_variables
-from pyfixest.estimation.internals.demean_ import DemeanCache, DemeanedData
-from pyfixest.estimation.internals.families import T_DIST, InferenceDist
 from pyfixest.estimation.internals.fit_ import fit_ols
 from pyfixest.estimation.internals.literals import (
     EstimatorKind,
     PredictionErrorOptions,
     PredictionType,
-    SolverOptions,
-    WeightsTypeOptions,
-    _validate_literal_argument,
 )
-from pyfixest.estimation.internals.model_state import (
-    ObservationWeights,
-    WithinLinearData,
+from pyfixest.estimation.models.base_regression_ import (
+    BaseRegression,
+    _check_vcov_input,
+    _deparse_vcov_input,
 )
-from pyfixest.estimation.internals.vcov_ import (
-    vcov_crv1,
-    vcov_crv3_fast,
-    vcov_hac,
-    vcov_hetero,
-    vcov_iid_ols,
-)
-from pyfixest.estimation.internals.vcov_utils import (
-    _compute_bread,
-    prepare_cluster_state,
-    run_crv_loop,
-)
-from pyfixest.estimation.models._result_accessor_mixin import ResultAccessorMixin
 from pyfixest.estimation.post_estimation.decomposition import (
     GelbachDecomposition,
     _decompose_arg_check,
 )
-from pyfixest.estimation.post_estimation.fixed_effects import (
-    FixedEffect,
-    build_fixed_effects,
-    check_fe_dtype_compatibility,
-    contrast_code_fixed_effects,
-    fixed_effects_to_frame,
-    predict_fixed_effects,
-    warn_on_unseen_fixed_effect_levels,
-)
-from pyfixest.estimation.post_estimation.prediction import _compute_prediction_error
-from pyfixest.estimation.post_estimation.wald import _wald_statistic
-from pyfixest.utils.dev_utils import (
-    DataFrameType,
-    _narwhals_to_pandas,
-)
-from pyfixest.utils.utils import (
-    capture_context,
-    get_ssc,
-)
+
+# Re-exported for backward compatibility: `pyfixest.estimation.feols_` has
+# always published the covariance-input helpers and the prediction literals
+# from this module.
+__all__ = [
+    "Feols",
+    "PredictionErrorOptions",
+    "PredictionType",
+    "_check_vcov_input",
+    "_deparse_vcov_input",
+    "decomposition_type",
+    "prediction_type",
+]
 
 decomposition_type = Literal["gelbach"]
 prediction_type = Literal["response", "link"]
 
 
-class Feols(ResultAccessorMixin):
+class Feols(BaseRegression):
     """
     Non user-facing class to estimate a linear regression via OLS.
 
@@ -108,166 +65,28 @@ class Feols(ResultAccessorMixin):
     ``DemeanCache``; the estimation runner supplies that cache for reuse across
     multiple fits.
 
-    Parameters
-    ----------
-    Y : np.ndarray
-        Dependent variable, a two-dimensional numpy array.
-    X : np.ndarray
-        Independent variables, a two-dimensional numpy array.
-    weights : np.ndarray
-        Weights, a one-dimensional numpy array.
-    collin_tol : float
-        Tolerance level for collinearity checks.
-    coefnames : list[str]
-        Names of the coefficients (of the design matrix X).
-    weights_name : Optional[str]
-        Name of the weights variable.
-    weights_type : Optional[str]
-        Type of the weights variable. Either "aweights" for analytic weights or
-        "fweights" for frequency weights.
-    solver : str, optional.
-        The solver to use for the regression. Can be "np.linalg.lstsq",
-        "np.linalg.solve", "scipy.linalg.solve" and "scipy.sparse.linalg.lsqr".
-        Defaults to "scipy.linalg.solve".
-    context : int or Mapping[str, Any]
-        A dictionary containing additional context variables to be used by
-        formulaic during the creation of the model matrix. This can include
-        custom factorization functions, transformations, or any other
-        variables that need to be available in the formula environment.
+    `Feols` is the ordinary-least-squares leaf of
+    [BaseRegression](/reference/estimation.models.base_regression_.BaseRegression.qmd).
+    It adds the OLS solve, the goodness-of-fit and Wald statistics reported for
+    a linear fit, the closed-form CRV3 jackknife, and the post-estimation
+    methods whose derivation assumes a single linear estimating equation:
+    `wildboottest()`, `ccv()`, `decompose()`, `update()`, `evalue()`, and
+    `pvalue_savi()`. Everything else is shared with the GLM and quantile
+    estimators and documented on the base class.
 
-    Attributes
-    ----------
-    _method : str
-        Specifies the method used for regression, set to "feols".
-    _is_iv : bool
-        Indicates whether instrumental variables are used, initialized as False.
+    See the base class for the constructor arguments and the retained
+    attributes. Difference-in-differences entry points return `Feols` results
+    and relabel `_method`; `test_treatment_heterogeneity()`, `aggregate()`, and
+    `iplot_aggregate()` are bound by those entry points and raise otherwise.
 
-    _Y : np.ndarray
-        The demeaned dependent variable, a two-dimensional numpy array.
-    _X : np.ndarray
-        The demeaned independent variables, a two-dimensional numpy array.
-    _X_is_empty : bool
-        Indicates whether the X array is empty.
-    _collin_tol : float
-        Tolerance level for collinearity checks.
-    _coefnames : list
-        Names of the coefficients (of the design matrix X).
-    _collin_vars : list
-        Variables identified as collinear.
-    _collin_index : list
-        Indices of collinear variables.
-    _Z : np.ndarray
-        Alias for the _X array, used for calculations.
-    _solver: str
-        The solver used for the regression.
-    _weights : np.ndarray
-        Array of weights for each observation.
-    _N : int
-        Number of observations.
-    _k : int
-        Number of independent variables (or features).
-    _data : Any
-        Data used in the regression, to be enriched outside of the class.
-    _fml : Any
-        Formula used in the regression, to be enriched outside of the class.
-    _has_fixef : bool
-        Indicates whether fixed effects are used.
-    _fixef : Any
-        Fixed effects used in the regression.
-    _icovars : Any
-        Internal covariates, to be enriched outside of the class.
-    _ssc_dict : dict
-        dictionary for sum of squares and cross products matrices.
-    _tZX : np.ndarray
-        Transpose of Z multiplied by X, set in get_fit().
-    _tXZ : np.ndarray
-        Transpose of X multiplied by Z, set in get_fit().
-    _tZy : np.ndarray
-        Transpose of Z multiplied by Y, set in get_fit().
-    _tZZinv : np.ndarray
-        Inverse of the transpose of Z multiplied by Z, set in get_fit().
-    _beta_hat : np.ndarray
-        Estimated regression coefficients.
-    _Y_hat_link : np.ndarray
-        Prediction at the level of the explanatory variable, i.e., the linear predictor X @ beta.
-    _Y_hat_response : np.ndarray
-        Prediction at the level of the response variable, i.e., the expected predictor E(Y|X).
-    _u_hat : np.ndarray
-        Residuals of the regression model.
-    _scores : np.ndarray
-        Scores used in the regression analysis.
-    _hessian : np.ndarray
-        Hessian matrix used in the regression.
-    _bread : np.ndarray
-        Bread matrix, used in calculating the variance-covariance matrix.
-    _vcov_type : Any
-        Type of variance-covariance matrix used.
-    _vcov_type_detail : Any
-        Detailed specification of the variance-covariance matrix type.
-    _is_clustered : bool
-        Indicates if clustering is used in the variance-covariance calculation.
-    _clustervar : Any
-        Variable used for clustering in the variance-covariance calculation.
-    _G : Any
-        Group information used in clustering.
-    _ssc : Any
-        Sum of squares and cross products matrix.
-    _vcov : np.ndarray
-        Variance-covariance matrix of the estimated coefficients.
-    _se : np.ndarray
-        Standard errors of the estimated coefficients.
-    _tstat : np.ndarray
-        T-statistics of the estimated coefficients.
-    _pvalue : np.ndarray
-        P-values associated with the t-statistics.
-    _conf_int : np.ndarray
-        Confidence intervals for the estimated coefficients.
-    _F_stat : Any
-        F-statistic for the model, set in get_Ftest().
-    _fixef_coefficients : dict[str, pyfixest.estimation.post_estimation.fixed_effects.FixedEffect]
-        Fixed effect estimates grouped by fixed effect.
-    _alpha : pd.DataFrame
-        A DataFrame with the estimated fixed effects.
-    _sumFE : np.ndarray
-        Sum of all fixed effects for each observation.
-    _rmse : float
-        Root mean squared error of the model.
-    _r2 : float
-        R-squared value of the model.
-    _r2_within : float
-        R-squared value computed on demeaned dependent variable.
-    _adj_r2 : float
-        Adjusted R-squared value of the model.
-    _adj_r2_within : float
-        Adjusted R-squared value computed on demeaned dependent variable.
-    _solver: Literal["np.linalg.lstsq", "np.linalg.solve", "scipy.linalg.solve",
-        "scipy.sparse.linalg.lsqr"],
-        default is "scipy.linalg.solve". Solver to use for the estimation.
-    _data: pd.DataFrame
-        The data frame used in the estimation. Deleted if arguments `lean = True`
-        or `store_data = False`.
-    _model_name: str
-        The name of the model. Usually just the formula string. If split estimation is used,
-        the model name will include the split variable and value.
-    _model_name_plot: str
-        The name of the model used when plotting and summarizing models. Usually identical to
-        `_model_name`. This might be different when pf.summary() or pf.coefplot() are called
-        and models with identical _model_name attributes are passed. In this case,
-        the _model_name_plot attribute will be modified.
-    _quantile: Optional[float]
-        The quantile used for quantile regression. None if not a quantile regression.
+    Examples
+    --------
+    ```{python}
+    import pyfixest as pf
 
-    # special for did
-    _res_cohort_eventtime_dict: Optional[dict[str, Any]]
-    _yname: Optional[str]
-    _gname: Optional[str]
-    _tname: Optional[str]
-    _idname: Optional[str]
-    _att: Optional[Any]
-    test_treatment_heterogeneity: Callable[..., Any]
-    aggregate: Callable[..., Any]
-    iplot_aggregate: Callable[..., Any]
-
+    fit = pf.feols("Y ~ X1 + X2 | f1", pf.get_data())
+    fit.tidy()
+    ```
     """
 
     _estimator: ClassVar[EstimatorKind] = "feols"
@@ -288,148 +107,15 @@ class Feols(ResultAccessorMixin):
         savi=unless(DIFFERENCE_IN_DIFFERENCES, FIXED_EFFECTS, WEIGHTED),
     )
 
-    def __init__(
-        self,
-        FixestFormula: FixestFormula,
-        data: pd.DataFrame,
-        ssc_dict: dict[str, str | bool],
-        drop_singletons: bool,
-        drop_intercept: bool,
-        weights: str | None,
-        weights_type: str | None,
-        collin_tol: float,
-        lookup_demeaned_data: dict[frozenset[int], DemeanedData],
-        solver: SolverOptions = "np.linalg.solve",
-        demeaner: AnyDemeaner | None = None,
-        lookup_preconditioner: dict[frozenset[int], Preconditioner] | None = None,
-        store_data: bool = True,
-        copy_data: bool = True,
-        lean: bool = False,
-        context: int | Mapping[str, Any] = 0,
-        sample_split_var: str | None = None,
-        sample_split_value: str | int | float | _AllSampleSentinel | None = None,
-    ) -> None:
-        self._sample_split_value = sample_split_value
-        self._sample_split_var = sample_split_var
-        self._model_name = (
-            FixestFormula.formula
-            if self._sample_split_var is None
-            else f"{FixestFormula.formula} (Sample: {self._sample_split_var} = {self._sample_split_value})"
-        )
-        self._model_name_plot = self._model_name
-        self._method = "feols"
-        self._is_iv = False
-        self._inference_dist: InferenceDist = T_DIST
-        self.FixestFormula = FixestFormula
+    def _bind_estimator_methods(self) -> None:
+        """Bind the difference-in-differences placeholders.
 
-        if self._sample_split_var is None:
-            pass
-        elif self._sample_split_value is _ALL_SAMPLE:
-            data = data.loc[data[sample_split_var].notnull()]
-        else:
-            data = data.loc[data[self._sample_split_var] == sample_split_value]
+        `event_study()` and `did2s()` return `Feols` results and rebind these
+        names on the result they hand back. A vanilla `feols()` fit keeps the
+        placeholder, which names the estimator the method belongs to.
+        """
 
-        data = data.reset_index(drop=True)
-
-        self._data = data.copy() if copy_data else data
-        self._ssc_dict = ssc_dict
-        self._drop_singletons = drop_singletons
-        self._drop_intercept = drop_intercept
-        self._weights_name = weights
-        self._weights_type = weights_type
-        self._has_weights = weights is not None
-        self._offset_name: str | None = None
-        self._offset: np.ndarray | None = None
-        self._collin_tol = collin_tol
-        self._solver = solver
-        if demeaner is None:
-            demeaner = MapDemeaner()
-        self._demeaner = demeaner
-        if isinstance(demeaner, LsmrDemeaner):
-            self._fixef_tol = max(demeaner.fixef_atol, demeaner.fixef_btol)
-        else:
-            self._fixef_tol = demeaner.fixef_tol
-        self._fixef_maxiter = demeaner.fixef_maxiter
-        self._demean_cache = DemeanCache(lookup_demeaned_data, lookup_preconditioner)
-        self._store_data = store_data
-        self._copy_data = copy_data
-        self._lean = lean
-        # Storage cleanup runs at the very end of the fit, so `lean` alone does
-        # not say whether the fit arrays are still there. Estimation itself goes
-        # through the same guarded methods.
-        self._fit_state_discarded = False
-        self._context = capture_context(context)
-
-        # attributes that have to be enriched outside of the class -
-        # not really optimal code change later
-        self._fml = FixestFormula.formula
-        self._has_fixef = False
-        self._fixef = (
-            str(FixestFormula.fixed_effects).replace(" ", "")
-            if FixestFormula.is_fixed_effects
-            else None
-        )
-        # self._coefnames = None
-        self._icovars = None
-
-        # set in get_fit()
-        self._tZX = np.array([])
-        # self._tZXinv = None
-        self._tXZ = np.array([])
-        self._tZy = np.array([])
-        self._tZZinv = np.array([])
-        self._beta_hat = np.array([])
-        self._Y_hat_link = np.array([])
-        self._Y_hat_response = np.array([])
-        self._u_hat = np.array([])
-        self._scores = np.array([])
-        self._hessian = np.array([])
-        self._bread = np.array([])
-
-        # set in vcov()
-        self._vcov_type = ""
-        self._vcov_type_detail = ""
-        self._is_clustered = False
-        self._clustervar: list[str] = []
-        self._G: list[int] = []
-        self._ssc = np.array([], dtype=np.float64)
-        self._vcov = np.array([])
-        self.n_separation_na = 0
-
-        # set in get_inference()
-        self._se = np.array([])
-        self._tstat = np.array([])
-        self._pvalue = np.array([])
-        self._conf_int = np.array([])
-
-        # set in fixef()
-        self._fixef_coefficients: dict[str, FixedEffect] = {}
-        self._alpha = None
-        self._sumFE = None
-
-        # set in get_performance()
-        self._rmse = np.nan
-        self._r2 = np.nan
-        self._r2_within = np.nan
-        self._adj_r2 = np.nan
-        self._adj_r2_within = np.nan
-
-        # special for poisson / glm
-        self.deviance: float | None = None
-
-        # special for did
-        self._res_cohort_eventtime_dict: dict[str, Any] | None = None
-        self._yname: str | None = None
-        self._gname: str | None = None
-        self._tname: str | None = None
-        self._idname: str | None = None
-        self._att: bool | None = None
-
-        # set functions inherited from other modules
-        self._bind_report_methods()
-
-        # DiD methods - assign placeholder functions
-        def _not_implemented_did(*args, **kwargs):
+        def _not_implemented_did(*args: Any, **kwargs: Any) -> Any:
             raise NotImplementedError(
                 "This method is only available for DiD models, not for vanilla 'feols'."
             )
@@ -437,187 +123,6 @@ class Feols(ResultAccessorMixin):
         self.test_treatment_heterogeneity = _not_implemented_did
         self.aggregate = _not_implemented_did
         self.iplot_aggregate = _not_implemented_did
-
-    def prepare_model_matrix(self):
-        """Build and retain the canonical formula-derived estimator inputs."""
-        model_matrix = model_matrix_fixest.create_model_matrix(
-            formula=self.FixestFormula,
-            data=self._data,
-            drop_singletons=self._drop_singletons,
-            drop_intercept=self._drop_intercept,
-            weights=self._weights_name,
-            offset=self._offset_name,
-            context=self._context,
-        )
-        self._publish_model_matrix(model_matrix)
-
-        # an empty drop still rebuilds the whole frame, so guard it
-        if self._na_index:
-            self._data.drop(index=list(self._na_index), inplace=True)
-
-        return model_matrix
-
-    # Deliberately untyped: an annotation makes mypy check this body, whose published
-    # attribute types are still transitional or loose; see the typing follow-up.
-    def _publish_model_matrix(self, model_matrix):
-        """Publish structurally immutable formula and observation-weight state."""
-        self._model_matrix = model_matrix
-        self._response = model_matrix.dependent.to_numpy(dtype=np.float64).flatten()
-        self._fe = model_matrix.fixed_effects
-        self._na_index = model_matrix.na_index
-        # TODO: set dynamically based on naming set in pyfixest.estimation.formula.factor_interaction._encode_i
-        independent = model_matrix.independent
-        is_icovar = (
-            independent.columns.str.contains(r"^.+::.+$")
-            if not independent.empty
-            else None
-        )
-        self._icovars = (
-            independent.columns[is_icovar].tolist()
-            if is_icovar is not None and is_icovar.any()
-            else None
-        )
-        self._X_is_empty = independent.shape[1] == 0
-        self._model_spec = model_matrix.model_spec
-
-        self._coefnames = independent.columns.tolist()
-        self._coefnames_z = (
-            model_matrix.instruments.columns.tolist()
-            if model_matrix.instruments is not None
-            else None
-        )
-        self._depvar = model_matrix.dependent.columns[0]
-
-        self._has_fixef = self._fe is not None
-        self._fixef = (
-            str(self.FixestFormula.fixed_effects).replace(" ", "")
-            if self.FixestFormula.is_fixed_effects
-            else None
-        )
-
-        self._k_fe = self._fe.nunique(axis=0) if self._has_fixef else None
-        self._n_fe = len(self._k_fe) if self._has_fixef else 0
-
-        self._observation_weights = self._set_observation_weights()
-        self._N = self._observation_weights.n_effective
-        self._N_rows = self._observation_weights.n_rows
-
-    def _validate_response(self) -> None:
-        """Validate estimator-specific response constraints, if any."""
-
-    def _set_observation_weights(self) -> ObservationWeights:
-        """Build canonical user-scale observation weights for this row sample."""
-        n_rows = len(self._model_matrix.dependent)
-        if self._model_matrix.weights is None:
-            return ObservationWeights.unweighted(n_rows=n_rows)
-
-        assert self._weights_type in ("aweights", "fweights")
-        weights_kind = cast(WeightsTypeOptions, self._weights_type)
-        return ObservationWeights.from_values(
-            self._model_matrix.weights.to_numpy().reshape(-1),
-            kind=weights_kind,
-        )
-
-    def _prepare_within_data(self) -> WithinLinearData:
-        """Return fixed-effect-residualized arrays in their original units."""
-        response_frame = self._model_matrix.dependent
-        design_frame = self._model_matrix.independent
-        response = response_frame.to_numpy(dtype=np.float64)
-        design = design_frame.to_numpy(dtype=np.float64)
-
-        if self._model_matrix.fixed_effects is not None:
-            response, design, _ = self._demean_cache.demean_yx(
-                response,
-                design,
-                y_names=response_frame.columns,
-                x_names=design_frame.columns,
-                fe=self._model_matrix.fixed_effects.to_numpy(),
-                weights=self._observation_weights.values,
-                na_index=self._na_index,
-                demeaner=self._demeaner,
-            )
-
-        return WithinLinearData(response=response, design=design)
-
-    @property
-    def preconditioner(self) -> Preconditioner | None:
-        """The within preconditioner used during demeaning, if any.
-
-        ``None`` when no preconditioner participated in the solve —
-        ``preconditioner='off'``, single-FE designs (MAP fallback), or any
-        non-within backend. Otherwise the instance built on the first solve
-        for this model's row sample. Pass it back via
-        ``LsmrDemeaner(backend='within', preconditioner=...)`` to skip the
-        setup phase on a later fit over the same design.
-        """
-        self._require_fit_arrays("preconditioner", arrays="the demeaning caches")
-        return self._demean_cache.lookup_preconditioner.get(self._na_index)
-
-    def _drop_multicollinear_within_data(
-        self, within_data: WithinLinearData
-    ) -> WithinLinearData:
-        """Return within data after the established unweighted rank check."""
-        design = within_data.design
-        if design.shape[1] > 0:
-            (
-                design,
-                self._coefnames,
-                self._collin_vars,
-                self._collin_index,
-            ) = drop_multicollinear_variables(
-                design,
-                self._coefnames,
-                self._collin_tol,
-            )
-
-        return WithinLinearData(
-            response=within_data.response,
-            design=design,
-            instruments=within_data.instruments,
-            endogenous=within_data.endogenous,
-        )
-
-    def _set_within_data(self, within_data: WithinLinearData) -> None:
-        """Publish canonical within data for this fit."""
-        self._within_data = within_data
-        self._X_is_empty = within_data.design.shape[1] == 0
-        self._k = within_data.design.shape[1]
-
-    # Read-only views on the canonical state objects. They exist so that long
-    # established attribute names keep working, but there is exactly one stored
-    # representation: assigning or deleting them is a programming error.
-
-    @property
-    def _Y(self) -> NDArray[np.float64]:
-        """Within-scale dependent variable, in response units."""
-        return self._within_data.response
-
-    @property
-    def _X(self) -> NDArray[np.float64]:
-        """Within-scale design matrix, not premultiplied by weights."""
-        return self._within_data.design
-
-    @property
-    def _Z(self) -> NDArray[np.float64]:
-        """Within-scale instruments; the design itself outside IV models."""
-        within_data = self._within_data
-        return (
-            within_data.design
-            if within_data.instruments is None
-            else within_data.instruments
-        )
-
-    @property
-    def _weights(self) -> NDArray[np.float64]:
-        """Observation weights as a column vector.
-
-        Unweighted fits store no weight vector, so the ones column is built on
-        access rather than kept alive for the lifetime of the result.
-        """
-        values = self._observation_weights.values
-        if values is None:
-            return np.ones((self._observation_weights.n_rows, 1), dtype=np.float64)
-        return values.reshape((-1, 1))
 
     def _get_predictors(self) -> None:
         self._Y_hat_link = self._response - self.resid()
@@ -663,404 +168,6 @@ class Feols(ResultAccessorMixin):
             self.get_performance()
             self.wald_test()
 
-    def _iter_fitted_models(self) -> tuple[Feols, ...]:
-        """Yield this fitted result to the result container."""
-        return (self,)
-
-    @property
-    def _fit_features(self) -> FitFeatures:
-        """Fit-level properties the capability rules read.
-
-        Rebuilt on every access: `vcov()` and the storage options mutate a
-        fitted result in place, so a cached snapshot could outlive the state it
-        describes.
-        """
-        family = getattr(self, "_family", None)
-        return FitFeatures(
-            estimator=self._estimator,
-            family=None if family is None else family.name,
-            is_iv=self._is_iv,
-            has_fixef=self._has_fixef,
-            has_weights=self._has_weights,
-            weights_kind=(
-                cast(WeightsTypeOptions, self._weights_type)
-                if self._has_weights
-                else None
-            ),
-            is_did=self._method in DID_METHOD_LABELS,
-        )
-
-    def _require_support(self, feature: Feature, *, subject: str | None = None) -> None:
-        """Reject a feature this fit does not support.
-
-        Parameters
-        ----------
-        feature : Feature
-            The feature the caller is about to run.
-        subject : str, optional
-            How the error message names the feature. Defaults to
-            `"<feature>()"`.
-        """
-        require_support(
-            capabilities=self._capabilities,
-            feature=feature,
-            features=self._fit_features,
-            subject=f"{feature}()" if subject is None else subject,
-        )
-
-    def _require_fit_arrays(
-        self,
-        method: str,
-        *,
-        arrays: str,
-        remedy: str = "Refit with lean=False.",
-    ) -> None:
-        """Reject a call whose input arrays `lean=True` discarded."""
-        if self._lean and self._fit_state_discarded:
-            raise RuntimeError(
-                f"{method}() is unavailable after fitting with lean=True because "
-                f"{arrays} were discarded. {remedy}"
-            )
-
-    def _require_estimation_data(
-        self,
-        method: str,
-        *,
-        remedy: str = "Refit with store_data=True.",
-    ) -> None:
-        """Reject a call whose estimation data the storage options discarded."""
-        if not hasattr(self, "_data"):
-            raise RuntimeError(
-                f"{method}() is unavailable when store_data=False or lean=True "
-                f"because the estimation data were discarded. {remedy}"
-            )
-
-    def vcov(
-        self,
-        vcov: str | dict[str, str],
-        vcov_kwargs: dict[str, str | int] | None = None,
-        data: DataFrameType | None = None,
-    ) -> Feols:
-        """
-        Compute covariance matrices for an estimated regression model.
-
-        Parameters
-        ----------
-        vcov : Union[str, dict[str, str]]
-            A string or dictionary specifying the type of variance-covariance matrix
-            to use for inference.
-            If a string, it can be one of "iid", "hetero", "HC1", "HC2", "HC3", "NW", "DK".
-            If a dictionary, it should have the format {"CRV1": "clustervar"} for
-            CRV1 inference or {"CRV3": "clustervar"}
-            for CRV3 inference. Note that CRV3 inference is currently not supported
-            for IV estimation.
-        vcov_kwargs : Optional[dict[str, any]]
-             Additional keyword arguments for the variance-covariance matrix.
-        data: Optional[DataFrameType], optional
-            The already-filtered estimation sample in its original estimation
-            order. This is required for data-dependent covariance updates when
-            the fitted model does not retain its input data. If None, tries to
-            fetch the data from the model object. Defaults to None.
-
-
-        Returns
-        -------
-        Feols
-            An instance of class [Feols](/reference/estimation.models.feols_.Feols.qmd) with updated inference.
-
-        Examples
-        --------
-        Updates the variance estimator of a fitted model without refitting it.
-        The model is modified in place and returned.
-
-        ```{python}
-        import pyfixest as pf
-
-        fit = pf.feols("Y ~ X1 + X2 | f1", pf.get_data())
-        fit.vcov("iid").tidy()
-        ```
-
-        ```{python}
-        # switch to cluster-robust inference
-        fit.vcov({"CRV1": "f1"}).tidy()
-        ```
-
-        See [On Small Sample Corrections](/explanation/ssc.qmd) for how the
-        `ssc` adjustments interact with each estimator.
-        """
-        self._require_fit_arrays(
-            "vcov",
-            arrays="the required estimation arrays",
-            remedy="Set vcov at estimation time or refit with lean=False.",
-        )
-
-        data_to_check: pd.DataFrame | None
-        if data is None:
-            data_to_check = getattr(self, "_data", None)
-        else:
-            try:
-                data_to_check = _narwhals_to_pandas(data)
-            except TypeError as e:
-                raise TypeError(
-                    f"The data set must be a DataFrame type. Received: {type(data)}"
-                ) from e
-            if len(data_to_check) != self._N_rows:
-                raise ValueError(
-                    "`data` passed to vcov() must contain exactly the already-filtered "
-                    "estimation sample in its original estimation order; expected "
-                    f"{self._N_rows} rows, received {len(data_to_check)}."
-                )
-
-        # assign estimated fixed effects, and fixed effects nested within cluster.
-
-        # deparse vcov input
-        _check_vcov_input(vcov=vcov, vcov_kwargs=vcov_kwargs, data=data_to_check)
-
-        vcov_type, vcov_type_detail, is_clustered, clustervar = _deparse_vcov_input(
-            vcov, self._has_fixef, self._is_iv
-        )
-
-        # Reject before any inference state is overwritten, so a failed update
-        # leaves the previous covariance estimate intact.
-        if vcov_type in {"HAC", "CRV"} and data_to_check is None:
-            self._require_estimation_data(
-                "vcov",
-                remedy="Pass the estimation sample via data= or refit with store_data=True.",
-            )
-
-        self._vcov_type = vcov_type
-        self._vcov_type_detail = vcov_type_detail
-        self._is_clustered = is_clustered
-        self._clustervar = clustervar
-
-        self._bread = _compute_bread(
-            self._is_iv, self._tXZ, self._tZZinv, self._tZX, self._hessian
-        )
-
-        if self._vcov_type == "iid":
-            self._ssc, self._df_k, self._df_t = get_ssc(
-                **self._make_ssc_kwargs(vcov_type="iid", G=1)
-            )
-            self._vcov = self._ssc * self._vcov_iid()
-
-        elif self._vcov_type == "hetero":
-            # fixest:::vcov_hetero_internal: adj = ifelse(ssc$cluster.adj, n/(n - 1), 1)
-            self._ssc, self._df_k, self._df_t = get_ssc(
-                **self._make_ssc_kwargs(vcov_type="hetero", G=self._N)
-            )
-            self._vcov = self._ssc * self._vcov_hetero()
-
-        elif self._vcov_type == "HAC":
-            assert data_to_check is not None
-            kw = vcov_kwargs or {}
-            self._lag = kw.get("lag")
-            self._time_id = kw.get("time_id")
-            self._panel_id = kw.get("panel_id")
-            self._ssc, self._df_k, self._df_t = get_ssc(
-                **self._make_ssc_kwargs(
-                    vcov_type="HAC",
-                    G=np.unique(data_to_check[self._time_id]).shape[0],
-                )  # number of unique time periods T used
-            )
-            self._vcov = self._ssc * self._vcov_hac(data=data_to_check)
-
-        elif self._vcov_type == "nid":
-            self._ssc, self._df_k, self._df_t = get_ssc(
-                **self._make_ssc_kwargs(vcov_type="hetero", G=self._N)
-            )
-            self._vcov = self._ssc * self._vcov_nid()
-
-        elif self._vcov_type == "CRV":
-            assert data_to_check is not None
-            prep = prepare_cluster_state(
-                data=data_to_check,
-                clustervar=self._clustervar,
-                ssc_dict=self._ssc_dict,
-                fixef=self._fixef,
-                fe=self._fe,
-                k_fe=self._k_fe,
-            )
-            self._cluster_df = prep.cluster_df
-            self._G = prep.G
-            self._vcov, self._ssc, self._df_k, self._df_t = run_crv_loop(
-                prep=prep,
-                k=self._k,
-                make_ssc_kwargs=self._make_ssc_kwargs,
-                cluster_vcov=partial(self._vcov_crv_cluster, data=data_to_check),
-            )
-        # update p-value, t-stat, standard error, confint
-        self.get_inference()
-
-        return self
-
-    def _make_ssc_kwargs(
-        self,
-        *,
-        vcov_type: str,
-        G: int | float | list[int],
-        vcov_sign: int = 1,
-        k_fe_nested: int = 0,
-        n_fe_fully_nested: int = 0,
-    ) -> dict:
-        "Bundle model-level and vcov-type-specific args for get_ssc()."
-        return {
-            "ssc_dict": self._ssc_dict,
-            "N": self._N,
-            "k": self._k,
-            "k_fe": self._k_fe.sum() if self._has_fixef else 0,
-            "n_fe": self._n_fe,
-            "vcov_type": vcov_type,
-            "G": G,
-            "vcov_sign": vcov_sign,
-            "k_fe_nested": k_fe_nested,
-            "n_fe_fully_nested": n_fe_fully_nested,
-        }
-
-    def _vcov_crv_cluster(
-        self,
-        clustid: np.ndarray,
-        cluster_col: np.ndarray,
-        *,
-        data: pd.DataFrame,
-    ) -> np.ndarray:
-        "Pick CRV1 / CRV3-fast / CRV3-slow for one cluster column."
-        if self._vcov_type_detail == "CRV1":
-            return self._vcov_crv1(clustid=clustid, cluster_col=cluster_col)
-
-        self._require_support("crv3", subject="CRV3 inference")
-        use_fast = not self._has_fixef and self._method == "feols" and not self._is_iv
-        if use_fast:
-            return self._vcov_crv3_fast(clustid=clustid, cluster_col=cluster_col)
-        return self._vcov_crv3_slow(
-            clustid=clustid,
-            cluster_col=cluster_col,
-            data=data,
-        )
-
-    def _vcov_iid(self):
-        return vcov_iid_ols(
-            residuals=self._u_hat,
-            bread=self._bread,
-            N=self._N,
-            weights=self._observation_weights.values,
-        )
-
-    def _leverage_weights(self) -> np.ndarray | None:
-        """Return weights used by the fitted normal equations."""
-        return self._observation_weights.values
-
-    def _fixef_weights(self) -> np.ndarray | None:
-        """Return weights used by fixed-effect coefficient recovery."""
-        return self._observation_weights.values
-
-    def _vcov_hetero(self):
-        observation_weights = self._observation_weights.values
-        return vcov_hetero(
-            scores=self._scores,
-            X=self._X,
-            tZX=self._tZX,
-            # Only the frequency-weight correction divides by the weights.
-            weights=(
-                observation_weights.reshape((-1, 1))
-                if observation_weights is not None and self._weights_type == "fweights"
-                else None
-            ),
-            leverage_weights=self._leverage_weights(),
-            weights_type=self._weights_type,
-            vcov_type_detail=self._vcov_type_detail,
-            bread=self._bread,
-            is_iv=self._is_iv,
-            tXZ=self._tXZ,
-            tZZinv=self._tZZinv,
-        )
-
-    def _vcov_hac(self, *, data: pd.DataFrame):
-        _time_id = self._time_id
-        _panel_id = self._panel_id
-
-        self._require_support("hac", subject="HAC inference (NW, DK)")
-
-        # some data checks on input pandas df
-        # time needs to be numeric or date else we cannot sort by time
-        if not np.issubdtype(data[_time_id], np.number) and not np.issubdtype(
-            data[_time_id], np.datetime64
-        ):
-            raise ValueError(
-                "The time variable must be numeric or date, else we cannot sort by time."
-            )
-
-        _time_arr = data[_time_id].to_numpy()
-        _panel_arr = data[_panel_id].to_numpy() if _panel_id is not None else None
-
-        return vcov_hac(
-            scores=self._scores,
-            time_arr=_time_arr,
-            panel_arr=_panel_arr,
-            lag=cast(int | None, self._lag),
-            vcov_type_detail=cast(Literal["NW", "DK"], self._vcov_type_detail),
-            bread=self._bread,
-            is_iv=self._is_iv,
-            tXZ=self._tXZ,
-            tZZinv=self._tZZinv,
-            tZX=self._tZX,
-        )
-
-    def _vcov_nid(self):
-        "Reject 'nid' covariance for estimators without a conditional density."
-        self._require_support("nid", subject="'nid' inference")
-        raise NotImplementedError(
-            f"'nid' inference is declared supported for models of type "
-            f"'{self._estimator}', but the class provides no implementation."
-        )
-
-    def _vcov_crv1(self, clustid: np.ndarray, cluster_col: np.ndarray):
-        return vcov_crv1(
-            scores=self._scores,
-            clustid=clustid,
-            cluster_col=cluster_col,
-            bread=self._bread,
-            is_iv=self._is_iv,
-            tXZ=self._tXZ,
-            tZZinv=self._tZZinv,
-            tZX=self._tZX,
-        )
-
-    def _vcov_crv3_fast(self, clustid, cluster_col):
-        return vcov_crv3_fast(
-            X=self._X,
-            Y=self._Y,
-            weights=self._observation_weights.values,
-            beta_hat=self._beta_hat,
-            clustid=clustid,
-            cluster_col=cluster_col,
-        )
-
-    def _estimation_refit_kwargs(self) -> dict[str, Any]:
-        """Return options needed to replay this estimator on modified data."""
-        demeaner = self._demeaner
-        if isinstance(demeaner, LsmrDemeaner) and isinstance(
-            demeaner.preconditioner, Preconditioner
-        ):
-            # A prebuilt factorization belongs to the original FE design. Keep
-            # its algorithmic variant, but rebuild it for the changed row set.
-            preconditioner = cast(
-                WithinPreconditionerName,
-                demeaner.preconditioner.variant.lower(),
-            )
-            demeaner = replace(demeaner, preconditioner=preconditioner)
-
-        return {
-            "weights": self._weights_name,
-            "weights_type": self._weights_type,
-            "ssc": dict(self._ssc_dict),
-            "fixef_rm": "singleton" if self._drop_singletons else "none",
-            "solver": self._solver,
-            "demeaner": demeaner,
-            "drop_intercept": self._drop_intercept,
-            "collin_tol": self._collin_tol,
-            "context": self._context,
-        }
-
     def _crv3_refit(self, data: pd.DataFrame) -> Feols:
         """Replay OLS for one leave-one-cluster-out sample."""
         # lazy loading to avoid circular import
@@ -1071,224 +178,6 @@ class Feols(ResultAccessorMixin):
             vcov="iid",
             **self._estimation_refit_kwargs(),
         )
-
-    def _vcov_crv3_slow(
-        self,
-        clustid: np.ndarray,
-        cluster_col: np.ndarray,
-        *,
-        data: pd.DataFrame,
-    ) -> np.ndarray:
-        beta_jack = np.zeros((len(clustid), self._k))
-
-        for ixg, g in enumerate(clustid):
-            # direct leave one cluster out implementation
-            refit_data = data[~np.equal(g, cluster_col)]
-            fit = self._crv3_refit(data=refit_data)
-            beta_jack[ixg, :] = fit.coef().to_numpy()
-
-        # optional: beta_bar in MNW (2022)
-        # center = "estimate"
-        # if center == 'estimate':
-        #    beta_center = beta_hat
-        # else:
-        #    beta_center = np.mean(beta_jack, axis = 0)
-        beta_center = self._beta_hat
-
-        vcov_mat = np.zeros((self._k, self._k))
-        for ixg, _ in enumerate(clustid):
-            beta_centered = beta_jack[ixg, :] - beta_center
-            vcov_mat += np.outer(beta_centered, beta_centered)
-
-        return vcov_mat
-
-    def _clear_attributes(self):
-        attributes = []
-
-        if not self._store_data:
-            attributes += ["_data", "_model_matrix"]
-
-        if self._lean:
-            # The array aliases are read-only views, so the state objects
-            # backing them are what has to go.
-            attributes += [
-                "_data",
-                "_model_matrix",
-                "_within_data",
-                "_observation_weights",
-                "_demean_cache",
-                "_fe",
-                "_response",
-                "_cluster_df",
-                "_tXZ",
-                "_tZy",
-                "_tZX",
-                "_tZZinv",
-                "_scores",
-                "_u_hat",
-                "_Y_hat_link",
-                "_Y_hat_response",
-            ]
-
-        for attr in attributes:
-            if hasattr(self, attr):
-                delattr(self, attr)
-
-        if self._lean:
-            self._fit_state_discarded = True
-
-    def capabilities(self) -> pd.DataFrame:
-        """
-        Report which post-estimation methods this fitted model supports.
-
-        Support depends on the estimator and on the fit: weights, absorbed
-        fixed effects, and instruments each withdraw methods whose derivation
-        does not cover them. Calling an unsupported method raises an error
-        carrying the same reason.
-
-        Returns
-        -------
-        pandas.DataFrame
-            Indexed by feature, with a boolean `supported` column and a
-            `reason` column that is None wherever the feature is available.
-            See [Which Methods Does Each Estimator Support?](/how-to/supported-methods.qmd)
-            for what each feature name refers to.
-
-        Notes
-        -----
-        Restrictions that depend on the arguments rather than on the fit, such
-        as the covariance types `evalue()` accepts or `update(inplace=True)`,
-        are documented with the individual methods.
-        [`pf.estimation.support_matrix()`](/reference/estimation.capabilities.support_matrix.qmd)
-        gives the same table across all estimators.
-
-        Examples
-        --------
-        ```{python}
-        import pyfixest as pf
-
-        fit = pf.feols("Y ~ X1 + X2", pf.get_data(), weights="weights")
-        fit.capabilities()
-        ```
-        """
-        reasons = self._capabilities.evaluate(self._fit_features)
-        index = pd.Index(list(reasons), name="feature")
-        return pd.DataFrame(
-            {
-                "supported": pd.Series(
-                    [reason is None for reason in reasons.values()], index=index
-                ),
-                # object dtype so a supported feature keeps its None reason
-                # instead of being coerced to a missing string value.
-                "reason": pd.Series(list(reasons.values()), index=index, dtype=object),
-            }
-        )
-
-    def wald_test(self, R=None, q=None, distribution="F"):
-        """
-        Conduct Wald test.
-
-        Compute a Wald test for a linear hypothesis of the form R * beta = q.
-        where R is m x k matrix, beta is a k x 1 vector of coefficients,
-        and q is m x 1 vector.
-        By default, tests the joint null hypothesis that all coefficients are zero.
-
-        This method producues the following attriutes
-
-        _dfd : int
-            degree of freedom in denominator
-        _dfn : int
-            degree of freedom in numerator
-        _wald_statistic : scalar
-            Wald-statistics computed for hypothesis testing
-        _f_statistic : scalar
-            Wald-statistics(when R is an indentity matrix, and q being zero vector)
-            computed for hypothesis testing
-        _p_value : scalar
-            corresponding p-value for statistics
-
-        Parameters
-        ----------
-        R : array-like, optional
-            The matrix R of the linear hypothesis.
-            If None, defaults to an identity matrix.
-        q : array-like, optional
-            The vector q of the linear hypothesis.
-            If None, defaults to a vector of zeros.
-        distribution : str, optional
-            The distribution to use for the p-value. Can be either "F" or "chi2".
-            Defaults to "F".
-
-        Returns
-        -------
-        pd.Series
-            A pd.Series with the Wald statistic and p-value.
-
-        Examples
-        --------
-        ```{python}
-        import numpy as np
-        import pandas as pd
-        import pyfixest as pf
-
-        data = pf.get_data()
-        fit = pf.feols("Y ~ X1 + X2| f1", data, vcov={"CRV1": "f1"}, ssc=pf.ssc(k_adj=False))
-
-        R = np.array([[1,-1]] )
-        q = np.array([0.0])
-
-        # Wald test
-        fit.wald_test(R=R, q=q, distribution = "chi2")
-        f_stat = fit._f_statistic
-        p_stat = fit._p_value
-
-        print(f"Python f_stat: {f_stat}")
-        print(f"Python p_stat: {p_stat}")
-        ```
-        """
-        k_fe = np.sum(self._k_fe.values) if self._has_fixef else 0
-
-        # If R is None, default to the identity matrix
-        R = np.eye(self._k) if R is None else np.atleast_2d(np.asarray(R, dtype=float))
-
-        W, self._dfn = _wald_statistic(
-            beta_hat=self._beta_hat,
-            vcov=self._vcov,
-            R=R,
-            q=q,
-        )
-
-        if self._is_clustered:
-            self._dfd = np.min(np.array(self._G)) - 1
-        else:
-            self._dfd = self._N - self._k - k_fe
-
-        self._wald_statistic = W
-
-        # The F distribution is only used for the joint test that all
-        # coefficients are zero (R identity, q zero).
-        if distribution == "F" and (
-            not np.array_equal(R, np.eye(self._k)) or (q is not None and np.any(q))
-        ):
-            warnings.warn(
-                "Distribution changed to chi2, as R is not an identity matrix and q is not a zero vector."
-            )
-            distribution = "chi2"
-
-        if distribution == "F":
-            self._f_statistic = W / self._dfn
-            self._p_value = 1 - f.cdf(self._f_statistic, dfn=self._dfn, dfd=self._dfd)
-            res = pd.Series({"statistic": self._f_statistic, "pvalue": self._p_value})
-        elif distribution == "chi2":
-            self._f_statistic = W / self._dfn
-            self._p_value = chi2.sf(self._wald_statistic, self._dfn)
-            res = pd.Series(
-                {"statistic": self._wald_statistic, "pvalue": self._p_value}
-            )
-        else:
-            raise ValueError("Distribution must be F or chi2")
-
-        return res
 
     def wildboottest(
         self,
@@ -1900,546 +789,6 @@ class Feols(ResultAccessorMixin):
 
         return med
 
-    def fixef(self, atol: float = 1e-06, btol: float = 1e-06) -> pd.DataFrame:
-        """
-        Compute the coefficients of (swept out) fixed effects for a regression model.
-
-        This method creates the following attributes:
-        - `_alpha` (pd.DataFrame): A DataFrame with the estimated fixed effects.
-        - `_sumFE` (np.array): An array with the sum of fixed effects for each
-        observation (i = 1, ..., N).
-
-        Parameters
-        ----------
-        atol : Float, default 1e-6
-            Stopping tolerance for scipy.sparse.linalg.lsqr().
-            See https://docs.scipy.org/doc/
-                scipy/reference/generated/scipy.sparse.linalg.lsqr.html
-        btol : Float, default 1e-6
-            Another stopping tolerance for scipy.sparse.linalg.lsqr().
-            See https://docs.scipy.org/doc/
-                scipy/reference/generated/scipy.sparse.linalg.lsqr.html
-
-        Returns
-        -------
-        pd.DataFrame
-            A tidy DataFrame with columns `variable`, `code`, `level`, and
-            `coefficient` containing the estimated fixed effects.
-
-        Examples
-        --------
-        ```{python}
-        import pyfixest as pf
-
-        fit = pf.feols("Y ~ X1 + X2 | f1", pf.get_data())
-        fixed_effects = fit.fixef()
-        fixed_effects.head()
-        ```
-        """
-        if not self._has_fixef:
-            raise ValueError("The regression model does not have fixed effects.")
-
-        self._require_support("fixef", subject="The fixef() method")
-
-        self._require_fit_arrays("fixef", arrays="the fitted arrays")
-        self._require_estimation_data("fixef")
-
-        fixef_weights = self._fixef_weights()
-
-        Y, X = self._model_spec[_ModelMatrixKey.main].get_model_matrix(
-            self._data,
-            output="pandas",
-            context=FORMULAIC_TRANSFORMS | {**self._context},
-        )
-        Y = Y.to_numpy().flatten().astype(np.float64)
-        if self._X_is_empty:
-            uhat = Y.flatten()
-        else:
-            # drop intercept, potentially multicollinear vars
-            X = X[self._coefnames].to_numpy()
-            if self._method == "fepois" or self._method.startswith("feglm"):
-                # determine residuals from estimated linear predictor
-                # equation (5.2) in Stammann (2018) http://arxiv.org/abs/1707.01815
-                Y = self._Y_hat_link
-                # _Y_hat_link contains the offset as part of eta; subtract it so
-                # that _sumFE represents the pure FE contribution and predict()
-                # can add the offset back from newdata without double-counting.
-                if self._offset_name is not None:
-                    assert self._offset is not None
-                    Y = Y - self._offset.flatten()
-            uhat = (Y - X @ self._beta_hat).flatten()
-        # one-hot encoding of fixed effects (treatment coding: reference level
-        # dropped for the second and subsequent FEs via ensure_full_rank=True).
-        contrast_coding = contrast_code_fixed_effects(
-            fixed_effects=self.FixestFormula.fixed_effects_wrapped,
-            fixed_effect_names=self._model_spec[
-                _ModelMatrixKey.fixed_effects
-            ].column_names,
-            data=self._data,
-            context=FORMULAIC_TRANSFORMS | {**self._context},
-            transform_state=self._model_spec[
-                _ModelMatrixKey.fixed_effects
-            ].transform_state,
-        )
-        fixed_effect_design = contrast_coding.matrix
-        solve_design = fixed_effect_design
-        if fixef_weights is not None:
-            weights_sqrt = np.sqrt(fixef_weights).flatten()
-            uhat *= weights_sqrt
-            weights_diag = diags(weights_sqrt, 0)
-            solve_design = weights_diag.dot(fixed_effect_design)
-
-        alpha = lsqr(solve_design, uhat, atol=atol, btol=btol)[0]
-
-        self._fixef_coefficients = build_fixed_effects(
-            fixed_effect_coefficients=alpha,
-            contrast_coding=contrast_coding,
-            transform_state=self._model_spec[
-                _ModelMatrixKey.fixed_effects
-            ].transform_state,
-        )
-        self._alpha = alpha
-        self._sumFE = fixed_effect_design.dot(alpha)
-
-        return fixed_effects_to_frame(self._fixef_coefficients)
-
-    def predict(
-        self,
-        newdata: DataFrameType | None = None,
-        atol: float = 1e-6,
-        btol: float = 1e-6,
-        type: PredictionType = "link",
-        se_fit: bool | None = False,
-        interval: PredictionErrorOptions | None = None,
-        alpha: float = 0.05,
-    ) -> np.ndarray | pd.DataFrame:
-        """
-        Predict values of the model on new data.
-
-        Return a flat np.array with predicted values of the regression model.
-        If new fixed effect levels are introduced in `newdata`, predicted values
-        for such observations will be set to NaN.
-
-        Parameters
-        ----------
-        newdata : DataFrameType, optional
-            A narwhals compatible DataFrame (polars, pandas, duckdb, etc).
-            If None (default), the data used for fitting the model is used.
-        type : str, optional
-            The type of prediction to be computed.
-            Can be either "response" (default) or "link". For linear models, both are
-            identical.
-        atol : Float, default 1e-6
-            Stopping tolerance for scipy.sparse.linalg.lsqr().
-            See https://docs.scipy.org/doc/
-                scipy/reference/generated/scipy.sparse.linalg.lsqr.html
-        btol : Float, default 1e-6
-            Another stopping tolerance for scipy.sparse.linalg.lsqr().
-            See https://docs.scipy.org/doc/
-                scipy/reference/generated/scipy.sparse.linalg.lsqr.html
-        type:
-            The type of prediction to be made. Can be either 'link' or 'response'.
-             Defaults to 'link'. 'link' and 'response' lead
-            to identical results for linear models.
-        se_fit: Optional[bool], optional
-            If True, the standard error of the prediction is computed. Only feasible
-            for models without fixed effects. GLMs are not supported. Defaults to False.
-        interval: str, optional
-            The type of interval to compute. Can be either 'prediction' or None.
-        alpha: float, optional
-            The alpha level for the confidence interval. Defaults to 0.05. Only
-            used if interval = "prediction" is not None.
-
-        Returns
-        -------
-        Union[np.ndarray, pd.DataFrame]
-            Returns a pd.Dataframe with columns "fit", "se_fit" and CIs if argument "interval=prediction".
-            Otherwise, returns a np.ndarray with the predicted values of the model or the prediction
-            standard errors if argument "se_fit=True".
-
-        Examples
-        --------
-        In-sample predictions:
-
-        ```{python}
-        import pyfixest as pf
-
-        data = pf.get_data()
-        fit = pf.feols("Y ~ X1 + X2 | f1", data)
-        fit.predict()[:5]
-        ```
-
-        Pass `newdata` to predict out of sample. Fixed effect levels that do not
-        appear in the estimation sample return missing values.
-
-        ```{python}
-        fit.predict(newdata=data.head())
-        ```
-        """
-        self._require_support("predict", subject="The predict() method")
-
-        if interval == "prediction" or se_fit:
-            self._require_support(
-                "prediction_errors", subject="Prediction with standard errors"
-            )
-
-        _validate_literal_argument(type, PredictionType)
-        if interval is not None:
-            _validate_literal_argument(interval, PredictionErrorOptions)
-
-        if newdata is None or se_fit or interval == "prediction":
-            self._require_fit_arrays(
-                "predict", arrays="the fitted design and residual arrays"
-            )
-
-        if newdata is None:
-            # note: no need to worry about fixed effects, as not supported with
-            # prediction errors; will throw error later;
-            X = self._X
-            y_hat = (
-                self._Y_hat_link
-                if type == "link" or self._method == "feols"
-                else self._Y_hat_response
-            )
-            n_observations = self._N_rows
-        else:
-            newdata = _narwhals_to_pandas(newdata).reset_index(drop=True)
-            n_observations = newdata.shape[0]
-            context = FORMULAIC_TRANSFORMS | {**self._context}
-            # Use na_action="drop" on each sub-spec separately because dependent variable
-            # may not be available in newdata, then intersect indices so a NaN in *any* variable
-            # (covariate or FE) marks the whole row as NaN in the output.
-            rhs_spec = self._model_spec[_ModelMatrixKey.main].rhs
-            X_mm, unseen = materialize_model_spec_with_unseen_mask(
-                rhs_spec, newdata, context
-            )
-            valid_idx = X_mm.index.to_numpy()
-            # rows with a categorical level unseen during fitting (in C()/i()) would
-            # be silently encoded as the reference level -> drop them to NaN instead,
-            # matching how unseen fixed-effect levels are handled below.
-            valid_idx = valid_idx[~unseen[valid_idx]]
-            if self._has_fixef:
-                # Fixed-effect levels are recovered from the estimation sample.
-                self._require_fit_arrays(
-                    "predict", arrays="the fitted design and residual arrays"
-                )
-                self._require_estimation_data("predict")
-                fe_spec = self._model_spec[_ModelMatrixKey.fixed_effects]
-                check_fe_dtype_compatibility(fe_spec, newdata)
-                # na_action="ignore" keeps unseen-level rows as NaN codes
-                fe_mm = fe_spec.get_model_matrix(
-                    newdata, context=context, na_action="ignore"
-                )
-                warn_on_unseen_fixed_effect_levels(fe_mm, fe_spec, newdata)
-                valid_fixed_effects = fe_mm.notna().all(axis="columns").to_numpy()
-                valid_idx = valid_idx[valid_fixed_effects[valid_idx]]
-                if self._sumFE is None:
-                    self.fixef(atol, btol)
-                fe_hat = predict_fixed_effects(
-                    model_matrix=fe_mm.loc[valid_idx],
-                    coefficients=self._fixef_coefficients,
-                )
-
-            X_coef = X_mm.loc[valid_idx, self._coefnames].to_numpy()
-            y_hat = np.full(n_observations, np.nan)
-            y_hat[valid_idx] = X_coef @ self._beta_hat
-            if self._has_fixef:
-                y_hat[valid_idx] += fe_hat
-            # Pad X to full size; NaN rows yield NaN SE/CI via einsum propagation.
-            X = np.full((n_observations, X_coef.shape[1]), np.nan)
-            X[valid_idx] = X_coef
-            if self._offset_name is not None:
-                offset_mm = self._model_spec[_ModelMatrixKey.offset].get_model_matrix(
-                    newdata,
-                    context=context,
-                    na_action="drop",
-                    output="pandas",
-                )
-                if not offset_mm.index.equals(newdata.index):
-                    raise ValueError(
-                        f"Offset expression '{self._offset_name}' evaluates to missing "
-                        "values in `newdata`."
-                    )
-
-                y_hat += offset_mm.iloc[:, 0].to_numpy()
-
-            if type == "response" and self._method == "fepois":
-                y_hat = np.exp(y_hat)
-
-        if se_fit or interval == "prediction":
-            prediction_df = _compute_prediction_error(
-                model=self,
-                nobs=n_observations,
-                yhat=y_hat,
-                X=X,
-                alpha=alpha,
-            )
-            if interval == "prediction":
-                return prediction_df
-            else:
-                return prediction_df["se_fit"].to_numpy()
-        else:
-            return y_hat
-
-    def ritest(
-        self,
-        resampvar: str,
-        cluster: str | None = None,
-        reps: int = 100,
-        type: str = "randomization-c",
-        rng: np.random.Generator | None = None,
-        choose_algorithm: str = "auto",
-        store_ritest_statistics: bool = False,
-        level: float = 0.95,
-    ) -> pd.Series:
-        """
-        Conduct Randomization Inference (RI) test against a null hypothesis of
-        `resampvar = 0`.
-
-        Parameters
-        ----------
-        resampvar : str
-            The name of the variable to be resampled.
-        cluster : str, optional
-            The name of the cluster variable in case of cluster random assignment.
-            If provided, `resampvar` is held constant within each `cluster`.
-            Defaults to None.
-        reps : int, optional
-            The number of randomization iterations. Defaults to 100.
-        type: str
-            The type of the randomization inference test.
-            Can be "randomization-c" or "randomization-t". Note that
-            the "randomization-c" is much faster, while the
-            "randomization-t" is recommended by Wu & Ding (JASA, 2021).
-        rng : np.random.Generator, optional
-            A random number generator. Defaults to None.
-        choose_algorithm: str, optional
-            The algorithm to use for the computation. Defaults to "auto".
-            The alternatives are "fast" and "slow". The fast algorithm requires
-            the optional `numba` extra (install via `pip install pyfixest[numba]`);
-            without it, the fast path raises an `ImportError`. The slow path
-            does not require numba.
-        include_plot: bool, optional
-            Whether to include a plot of the distribution p-values. Defaults to False.
-        store_ritest_statistics: bool, optional
-            Whether to store the simulated statistics of the RI procedure.
-            Defaults to False. If True, stores the simulated statistics
-            in the model object via the `ritest_statistics` attribute as a
-            numpy array.
-        level: float, optional
-            The level for the confidence interval of the randomization inference
-            p-value. Defaults to 0.95.
-
-        Returns
-        -------
-        A pd.Series with the regression coefficient of `resampvar` and the p-value
-        of the RI test. Additionally, reports the standard error and the confidence
-        interval of the p-value.
-
-        Examples
-        --------
-        ```{python}
-
-        #| echo: true
-        #| results: asis
-        #| include: true
-
-        import pyfixest as pf
-        data = pf.get_data()
-        fit = pf.feols("Y ~ X1 + X2", data=data)
-
-        # Conduct a randomization inference test for the coefficient of X1
-        fit.ritest("X1", reps=1000)
-
-        # use randomization-t instead of randomization-c
-        fit.ritest("X1", reps=1000, type="randomization-t")
-
-        # store statistics for plotting
-        fit.ritest("X1", reps=1000, store_ritest_statistics=True)
-        ```
-        """
-        from pyfixest.estimation.post_estimation.ritest import (
-            _HAS_NUMBA,
-            _decode_resampvar,
-            _get_ritest_pvalue,
-            _get_ritest_stats_fast,
-            _get_ritest_stats_slow,
-        )
-
-        resampvar = resampvar.replace(" ", "")
-        resampvar_, h0_value, hypothesis, test_type = _decode_resampvar(resampvar)
-
-        self._require_support("ritest", subject="Randomization inference")
-
-        # check that resampvar in _coefnames
-        if resampvar_ not in self._coefnames:
-            raise ValueError(f"{resampvar_} not found in the model's coefficients.")
-
-        self._require_fit_arrays("ritest", arrays="the fitted arrays")
-        self._require_estimation_data("ritest")
-
-        if cluster is not None and cluster not in self._data:
-            raise ValueError(f"The variable {cluster} is not found in the data.")
-
-        clustervar_arr = (
-            self._data[cluster].to_numpy().reshape(-1, 1) if cluster else None
-        )
-
-        if clustervar_arr is not None and np.any(np.isnan(clustervar_arr)):
-            raise ValueError(
-                """
-            The cluster variable contains missing values. This is not allowed
-            for randomization inference via `ritest()`.
-            """
-            )
-
-        # update vcov if cluster provided but not in model
-        if cluster is not None and not self._is_clustered:
-            warnings.warn(
-                "The initial model was not clustered. CRV1 inference is computed and stored in the model object."
-            )
-            self.vcov({"CRV1": cluster})
-
-        rng = np.random.default_rng() if rng is None else rng
-
-        sample_coef = np.array(self.coef().xs(resampvar_))
-        sample_tstat = np.array(self.tstat().xs(resampvar_))
-        sample_stat = sample_tstat if type == "randomization-t" else sample_coef
-
-        if type not in ["randomization-t", "randomization-c"]:
-            raise ValueError("type must be 'randomization-t' or 'randomization-c.")
-
-        # always run slow algorithm for randomization-t
-        choose_algorithm = "slow" if type == "randomization-t" else choose_algorithm
-
-        if choose_algorithm == "auto":
-            choose_algorithm = "fast" if _HAS_NUMBA else "slow"
-
-        assert isinstance(reps, int) and reps > 0, "reps must be a positive integer."
-
-        if choose_algorithm == "slow" or self._method == "fepois":
-            vcov_input: str | dict[str, str]
-            if cluster is not None:
-                vcov_input = {"CRV1": cluster}
-            else:
-                # "iid" for models without controls, else HC1
-                vcov_input = (
-                    "hetero"
-                    if (self._has_fixef and len(self._coefnames) > 1)
-                    or len(self._coefnames) > 2
-                    else "iid"
-                )
-
-            # for performance reasons
-            if type == "randomization-c":
-                vcov_input = "iid"
-
-            ri_stats = _get_ritest_stats_slow(
-                data=self._data,
-                resampvar=resampvar_,
-                clustervar_arr=clustervar_arr,
-                fml=self._fml,
-                reps=reps,
-                vcov=vcov_input,
-                type=type,
-                rng=rng,
-                model=self._method,
-                refit_kwargs=self._estimation_refit_kwargs(),
-            )
-
-        else:
-            observation_weights = self._observation_weights.values
-            # The demeaning kernel behind the fast path always needs a vector.
-            weights = (
-                np.ones(self._N_rows, dtype=np.float64)
-                if observation_weights is None
-                else observation_weights
-            )
-            fval_df = (
-                self._data[self._fixef.split("+")] if self._fixef is not None else None
-            )
-            D = self._data[resampvar_].to_numpy()
-
-            ri_stats = _get_ritest_stats_fast(
-                Y=self._Y,
-                X=self._X,
-                D=D,
-                coefnames=self._coefnames,
-                resampvar=resampvar_,
-                clustervar_arr=clustervar_arr,
-                reps=reps,
-                rng=rng,
-                fval_df=fval_df,
-                weights=weights,
-            )
-
-        ri_pvalue, se_pvalue, ci_pvalue = _get_ritest_pvalue(
-            sample_stat=sample_stat,
-            ri_stats=ri_stats[1:],
-            method=test_type,
-            h0_value=h0_value,
-            level=level,
-        )
-
-        if store_ritest_statistics:
-            self._ritest_statistics = ri_stats
-            self._ritest_pvalue = ri_pvalue
-            self._ritest_sample_stat = sample_stat - h0_value
-
-        res = pd.Series(
-            {
-                "H0": hypothesis,
-                "ri-type": type,
-                "Estimate": sample_coef,
-                "Pr(>|t|)": ri_pvalue,
-                "Std. Error (Pr(>|t|))": se_pvalue,
-            }
-        )
-
-        alpha = 1 - level
-        ci_lower_name = str(f"{alpha / 2 * 100:.1f}% (Pr(>|t|))")
-        ci_upper_name = str(f"{(1 - alpha / 2) * 100:.1f}% (Pr(>|t|))")
-        res[ci_lower_name] = ci_pvalue[0]
-        res[ci_upper_name] = ci_pvalue[1]
-
-        if cluster is not None:
-            res["Cluster"] = cluster
-
-        return res
-
-    def plot_ritest(self, plot_backend="lets_plot"):
-        """
-        Plot the distribution of the Randomization Inference Statistics.
-
-        Parameters
-        ----------
-        plot_backend : str, optional
-            The plotting backend to use. Defaults to "lets_plot". Alternatively,
-            "matplotlib" is available.
-
-        Returns
-        -------
-        A lets_plot or matplotlib figure with the distribution of the Randomization
-        Inference Statistics.
-        """
-        from pyfixest.estimation.post_estimation.ritest import _plot_ritest_pvalue
-
-        if not hasattr(self, "_ritest_statistics"):
-            raise ValueError(
-                """
-                            The randomization inference statistics have not been stored
-                            in the model object. Please set `store_ritest_statistics=True`
-                            when calling `ritest()`
-                            """
-            )
-
-        ri_stats = self._ritest_statistics
-        sample_stat = self._ritest_sample_stat
-
-        return _plot_ritest_pvalue(
-            ri_stats=ri_stats, sample_stat=sample_stat, plot_backend=plot_backend
-        )
-
     def update(
         self, X_new: np.ndarray, y_new: np.ndarray, inplace: bool = False
     ) -> np.ndarray:
@@ -2505,147 +854,3 @@ class Feols(ResultAccessorMixin):
         beta_n_plus_1 = self._beta_hat + gamma_n_plus_1 @ epsi_n_plus_1
 
         return beta_n_plus_1
-
-
-def _check_vcov_input(
-    vcov: str | dict[str, str],
-    vcov_kwargs: dict[str, Any] | None,
-    data: pd.DataFrame | None,
-):
-    """
-    Check the input for the vcov argument in the Feols class.
-
-    Parameters
-    ----------
-    vcov : Union[str, dict[str, str]]
-        The vcov argument passed to the Feols class.
-    vcov_kwargs : Optional[dict[str, Any]]
-        The vcov_kwargs argument passed to the Feols class.
-    data : pd.DataFrame or None
-        The estimation sample, or None when the fitted model retains none.
-
-    Returns
-    -------
-    None
-    """
-    assert isinstance(vcov, (dict, str, list)), "vcov must be a dict, string or list"
-    if isinstance(vcov, dict):
-        assert next(iter(vcov.keys())) in [
-            "CRV1",
-            "CRV3",
-        ], "vcov dict key must be CRV1 or CRV3"
-        assert isinstance(next(iter(vcov.values())), str), (
-            "vcov dict value must be a string"
-        )
-        deparse_vcov = next(iter(vcov.values())).split("+")
-        assert len(deparse_vcov) <= 2, "not more than twoway clustering is supported"
-
-    if isinstance(vcov, list):
-        assert all(isinstance(v, str) for v in vcov), "vcov list must contain strings"
-        if data is None:
-            raise RuntimeError(
-                "A vcov column list requires estimation data. Pass data= or fit "
-                "with store_data=True."
-            )
-        assert all(v in data.columns for v in vcov), (
-            "vcov list must contain columns in the data"
-        )
-    if isinstance(vcov, str):
-        assert vcov in [
-            "iid",
-            "hetero",
-            "HC1",
-            "HC2",
-            "HC3",
-            "NW",
-            "DK",
-            "nid",
-        ], (
-            "vcov string must be iid, hetero, HC1, HC2, HC3, NW, or DK, or for quantile regression, 'nid'."
-        )
-
-        # check that time_id is provided if vcov is NW or DK
-        if (
-            vcov in {"NW", "DK"}
-            and vcov_kwargs is not None
-            and "time_id" not in vcov_kwargs
-        ):
-            raise ValueError("Missing required 'time_id' for NW/DK vcov")
-
-
-def _deparse_vcov_input(vcov: str | dict[str, str], has_fixef: bool, is_iv: bool):
-    """
-    Deparse the vcov argument passed to the Feols class.
-
-    Parameters
-    ----------
-    vcov : Union[str, dict[str, str]]
-        The vcov argument passed to the Feols class.
-    has_fixef : bool
-        Whether the regression has fixed effects.
-    is_iv : bool
-        Whether the regression is an IV regression.
-
-    Returns
-    -------
-    vcov_type : str
-        The type of vcov to be used. Either "iid", "hetero", or "CRV".
-    vcov_type_detail : str or list
-        The type of vcov to be used, with more detail. Options include "iid",
-        "hetero", "HC1", "HC2", "HC3", "CRV1", or "CRV3".
-    is_clustered : bool
-        Indicates whether the vcov is clustered.
-    clustervar : str
-        The name of the cluster variable.
-    """
-    if isinstance(vcov, dict):
-        vcov_type_detail = next(iter(vcov.keys()))
-        deparse_vcov = next(iter(vcov.values())).split("+")
-        if isinstance(deparse_vcov, str):
-            deparse_vcov = [deparse_vcov]
-        deparse_vcov = [x.replace(" ", "") for x in deparse_vcov]
-    elif isinstance(vcov, (list, str)):
-        vcov_type_detail = vcov
-    else:
-        raise TypeError("arg vcov needs to be a dict, string or list")
-
-    if vcov_type_detail == "iid":
-        vcov_type = "iid"
-        is_clustered = False
-    elif vcov_type_detail in ["hetero", "HC1", "HC2", "HC3"]:
-        vcov_type = "hetero"
-        is_clustered = False
-        if vcov_type_detail in ["HC2", "HC3"]:
-            if has_fixef:
-                raise VcovTypeNotSupportedError(
-                    "HC2 and HC3 inference types are not supported for regressions with fixed effects."
-                )
-            if is_iv:
-                raise VcovTypeNotSupportedError(
-                    "HC2 and HC3 inference types are not supported for IV regressions."
-                )
-    elif vcov_type_detail in ["NW", "DK"]:
-        vcov_type = "HAC"
-        is_clustered = False
-
-    elif vcov_type_detail in ["CRV1", "CRV3"]:
-        vcov_type = "CRV"
-        is_clustered = True
-
-    elif vcov_type_detail == "nid":
-        vcov_type = "nid"
-        is_clustered = False
-
-    clustervar = deparse_vcov if is_clustered else None
-
-    # loop over clustervar to change "^" to "_"
-    if clustervar and "^" in clustervar:
-        clustervar = [x.replace("^", "_") for x in clustervar]
-        warnings.warn(
-            f"""
-            The '^' character in the cluster variable name is replaced by '_'.
-            In consequence, the clustering variable(s) is (are) named {clustervar}.
-            """
-        )
-
-    return vcov_type, vcov_type_detail, is_clustered, clustervar
